Refactored GrailsRequestHandler
Introduced ActionSpecification as a concept for different types of actions. Currently the default actions are modelled and we have a factory for each of the action specifications. This makes the action specifications itself unit testable.

As a result of this refactoring the expected json has also been modified to be correct with respect to the index operation.

There is still a broken test for /application/index but thats for a different commit.